Rename host_ir files to ir
