+++ csrc/host_ir/container.cpp
@@ -8,7 +8,7 @@
 
 #include <host_ir/container.h>
 
-#include <host_ir/host_ir.h>
+#include <host_ir/ir.h>
 #include <ir/builder.h>
 #include <ir/cloner.h>
 #include <ir/printer.h>
