+++ csrc/host_ir/container.h
@@ -10,7 +10,7 @@
 #include <list>
 
 #include <fusion.h>
-#include <host_ir/host_ir.h>
+#include <host_ir/ir.h>
 #include <ir/internal_nodes.h>
 #include <runtime/executor.h>
