+++ csrc/host_ir/ir.cpp
@@ -6,7 +6,7 @@
  */
 // clang-format on
 
-#include <host_ir/host_ir.h>
+#include <host_ir/ir.h>
 
 #include <algorithm>
 #include <memory>
