+++ tests/cpp/test_host_ir_stream_lowering.cpp
@@ -16,7 +16,7 @@
 #include <fusion.h>
 #include <host_ir/container.h>
 #include <host_ir/evaluator.h>
-#include <host_ir/host_ir.h>
+#include <host_ir/ir.h>
 #include <host_ir/lower.h>
 #include <host_ir/pass/stream_parallel_type.h>
 #include <ir/all_nodes.h>
